Add files via upload
Here is my portion of the project. I have some of the code blocks for the different image analysis programs I was trying to use, but I mentioned Anna found and used ImageR to get data set. I then ran a code block for an aov test of results and a simple plot. All that would be left is running the TukeyHSD code block since the p value was .0129